+++ src/components/form.js
@@ -11,9 +11,13 @@ export default React.createClass({
 
         // TODO: concretize type
         validationSchema: React.PropTypes.any.isRequired,
+        strategy: React.PropTypes.object.isRequired,
         formStateCursor: BaobabPropTypes.cursor,
         validateOnFly: React.PropTypes.bool,
         useHtmlForm: React.PropTypes.bool,
+
+        className: React.PropTypes.string,
+        style: React.PropTypes.object,
     },
 
     childContextTypes: {
@@ -24,6 +28,7 @@ export default React.createClass({
         form: React.PropTypes.object,
     },
 
+    subscribers: [],
     _isHtmlForm: null,
 
     getDefaultProps() {
@@ -51,6 +56,9 @@ export default React.createClass({
 
                 submit: this.submit,
                 validate: this.validate,
+
+                subscribe: this.subscribe,
+                unsubscribe: this.unsubscribe,
             },
         };
     },
@@ -63,6 +71,7 @@ export default React.createClass({
         return {
             errors: {},
             dirtyStates: {},
+            isFormDirty: false,
         };
     },
 
@@ -81,16 +90,21 @@ export default React.createClass({
     },
 
     render() {
+        const props = _.pick(this.props, ['className', 'style']);
+
         if (this.isHtmlForm()) {
             return (
-                <form noValidate {...this.props} onSubmit={this.onFormSubmit}>
+                <form
+                    noValidate
+                    {...props}
+                    onSubmit={this.onFormSubmit}>
                     {this.props.children}
                 </form>
             );
         }
 
         return (
-            <div {...this.props}>
+            <div {...props}>
                 {this.props.children}
             </div>
         );
@@ -116,11 +130,50 @@ export default React.createClass({
         return this._isHtmlForm;
     },
 
+    publish(...args) {
+        _.each(this.subscribers, (subscriber) => subscriber(...args));
+    },
+
+    subscribe(subscriber) {
+        this.subscribers = _.concat(this.subscribers, subscriber);
+
+        // Notify new subscriber about initial form state
+        subscriber(this.getFormState(), null);
+    },
+
+    unsubscribe(subscriber) {
+        this.subscribers = _.without(this.subscribers, subscriber);
+    },
+
+    onFormStateUpdate(data, previousData) {
+        // Notify subscribers about changed data only
+        if (_.isEqual(data, previousData)) {
+            return;
+        }
+
+        this.publish(data, previousData);
+    },
+
     setFormState(nextState) {
+        // It is necessary because validation strategy is asynchronous and validation's
+        // callback might be called after component unmount
+        if (!this.isMounted()) {
+            return;
+        }
+
+        const prevState = this.getFormState();
+
         if (this.props.formStateCursor) {
+            this.props.formStateCursor.once(
+                'update',
+                ({ data }) => this.onFormStateUpdate(data.currentData, prevState)
+            );
             this.props.formStateCursor.merge(nextState);
         } else {
-            this.setState(nextState);
+            this.setState(
+                nextState,
+                () => this.onFormStateUpdate(this.state, prevState)
+            );
         }
     },
 
@@ -147,8 +200,12 @@ export default React.createClass({
         return this.validate(this.props.onSubmit, this.props.onInvalidSubmit);
     },
 
+    getFormData() {
+        return this.props.cursor.get();
+    },
+
     validate(successCallback, errorCallback) {
-        const data = this.props.cursor.get();
+        const data = this.getFormData();
         const schema = _.isFunction(this.props.validationSchema) ?
             this.props.validationSchema(data) : this.props.validationSchema;
 
@@ -188,14 +245,17 @@ export default React.createClass({
 
     isDirty(fieldPath) {
         const formState = this.getFormState();
+        const isFormDirty = formState.isFormDirty;
+
+        if (isFormDirty) {
+            return true;
+        }
+
+        if (!fieldPath) {
+            return !_.isEmpty(formState.dirtyStates);
+        }
 
         return !!_.get(formState.dirtyStates, fieldPath);
-    },
-
-    resetDirtyStates() {
-        this.setFormState({
-            dirtyStates: {},
-        });
     },
 
     resetValidationErrors() {
@@ -209,10 +269,26 @@ export default React.createClass({
     },
 
     setDirtyState(fieldPath) {
+        if (!fieldPath) {
+            this.setFormState({
+                dirtyStates: {},
+                isFormDirty: true,
+            });
+            return;
+        }
+
         this.updateDirtyState(fieldPath, true);
     },
 
     setPristineState(fieldPath) {
+        if (!fieldPath) {
+            this.setFormState({
+                dirtyStates: {},
+                isFormDirty: false,
+            });
+            return;
+        }
+
         this.updateDirtyState(fieldPath, false);
     },
 
@@ -221,10 +297,13 @@ export default React.createClass({
             return;
         }
 
-        const formState = this.getFormState();
+        let { dirtyStates } = this.getFormState();
+
+        dirtyStates = _.cloneDeep(dirtyStates || {});
+        _.set(dirtyStates, fieldPath, dirtyState);
 
         this.setFormState({
-            dirtyStates: _.set(formState.dirtyStates || {}, fieldPath, dirtyState),
+            dirtyStates,
         });
     },
 });