+++ src/components/validation-error.js
@@ -1,8 +1,12 @@
 import React from 'react';
 import classNames from 'classnames';
+import PureRenderMixin from 'react-addons-pure-render-mixin';
+import { FormComponentMixin } from '../mixins';
 
 export default React.createClass({
     displayName: 'ValidationError',
+
+    mixins: [FormComponentMixin, PureRenderMixin],
 
     contextTypes: {
         form: React.PropTypes.object.isRequired,
@@ -14,12 +18,14 @@ export default React.createClass({
           React.PropTypes.array.isRequired,
         ]),
         className: React.PropTypes.string,
+        dirtyClassName: React.PropTypes.string,
         alwaysShow: React.PropTypes.bool,
     },
 
     getDefaultProps() {
         return {
             alwaysShow: false,
+            dirtyClassName: '_dirty',
         };
     },
 
@@ -31,21 +37,21 @@ export default React.createClass({
     },
 
     render() {
-        const form = this.context.form;
-
-        const error = form.getValidationErrors(this.props.fieldPath);
-        const isValid = !error;
-        const isDirty = form.isDirty(this.props.fieldPath);
-        const className = classNames(this.props.className, {
-            _dirty: isDirty,
+        const { alwaysShow, className, dirtyClassName } = this.props;
+        const errors = this.getErrors();
+        const isValid = this.isValid();
+        const isDirty = this.isDirty();
+        const generatedClassName = classNames(className, {
+            [dirtyClassName]: isDirty,
         });
-        if (isValid || !this.props.alwaysShow && !isDirty) {
+
+        if (isValid || !alwaysShow && !isDirty) {
             return null;
         }
 
         return (
-            <div className={className}>
-                {error}
+            <div className={generatedClassName}>
+                {errors}
             </div>
         );
     },