+++ test/components/check-box.spec.js
@@ -5,9 +5,9 @@ import Baobab from 'baobab';
 import SchemaBranchMixin from 'baobab-react-schemabranchmixin';
 import TestUtils from 'react-addons-test-utils';
 import yup from 'yup';
-import { Form, CheckBox } from '../../src/components';
+import sinon from 'sinon';
 import { Root } from '../utils';
-import sinon from 'imports?define=>false,require=>false!sinon/pkg/sinon-2.0.0-pre.js';
+import { Form, CheckBox } from '../../src/components';
 
 const tree = new Baobab(
     {},
@@ -84,7 +84,10 @@ describe('CheckBox', () => {
         TestUtils.Simulate.change(inputNode, { target: { checked: true } });
 
         clock.tick(1);
-        onChangeSpy.should.have.been.calledWith('checked', 'unchecked');
+        onChangeSpy.should.have.been.calledWithMatch(
+            sinon.match.object,
+            { value: 'checked', previousValue: 'unchecked' }
+        );
         tree.get('form', 'field').should.be.equal('checked');
         inputNode.checked.should.be.true;
     });
@@ -103,7 +106,10 @@ describe('CheckBox', () => {
         TestUtils.Simulate.change(inputNode, { target: { checked: false } });
 
         clock.tick(1);
-        onChangeSpy.should.have.been.calledWith('unchecked', 'checked');
+        onChangeSpy.should.have.been.calledWithMatch(
+            sinon.match.object,
+            { value: 'unchecked', previousValue: 'checked' }
+        );
         tree.get('form', 'field').should.be.equal('unchecked');
         inputNode.checked.should.be.false;
     });