+++ test/components/form.spec.js
@@ -2,11 +2,10 @@
 import React from 'react';
 import Baobab from 'baobab';
 import SchemaBranchMixin from 'baobab-react-schemabranchmixin';
-import BaobabPropTypes from 'baobab-prop-types';
 import TestUtils from 'react-addons-test-utils';
 import yup from 'yup';
-import { Form, Submit } from '../../src/components';
 import { Root } from '../utils';
+import { Form, Submit } from '../../src/components';
 
 const tree = new Baobab(
     {},
@@ -188,11 +187,18 @@ describe('Form without on fly validation', () => {
         expect(formComponent.isValid('second.field')).to.be.true;
     });
 
-    it('should resetDirtyStates works correctly', () => {
-        formComponent.resetDirtyStates();
+    it('should setDirtyState without arguments works correctly', () => {
+        formComponent.setDirtyState();
+        expect(formComponent.isDirty('firstName')).to.be.true;
+        expect(formComponent.isDirty()).to.be.true;
+    });
+
+    it('should setPristineState without arguments works correctly', () => {
+        formComponent.setPristineState();
         expect(formComponent.isDirty('firstName')).to.be.false;
         expect(formComponent.isDirty('lastName')).to.be.false;
         expect(formComponent.isDirty('second.field')).to.be.false;
+        expect(formComponent.isDirty()).to.be.false;
     });
 
     it('should setDirtyState works correctly', () => {